Pulled MultiSSH code into its own class
