+++ mRemoteV1/App/NativeMethods.cs
@@ -382,12 +382,12 @@ namespace mRemoteNG.App
         public const int WM_KEYUP = 0x101;
 
         /// <summary>
-        /// 
+        /// Posted to the window with the keyboard focus when a WM_KEYDOWN message is translated by the TranslateMessage function. The WM_CHAR message contains the character code of the key that was pressed.
         /// </summary>
         public const int WM_CHAR = 0x102;
 
         /// <summary>
-        /// 
+        /// Sent when the user selects a command item from a menu, when a control sends a notification message to its parent window, or when an accelerator keystroke is translated.
         /// </summary>
         public const int WM_COMMAND = 0x111;
